inference_demo_kafka/requirements.txt
spacy==3.0.6
setuptools
wheel
kafka-python
